LayoutTests/imported/w3c/web-platform-tests/trusted-types/block-string-assignment-to-DOMParser-parseFromString-expected.txt


PASS document.innerText assigned via policy (successful HTML transformation).
FAIL `document.innerText = string` throws. assert_throws_js: function "_ => {
      var doc = parser.parseFromString("Fail", "text/html");
    }" did not throw
FAIL 'document.innerText = null' throws assert_throws_js: function "_ => {
      var doc = parser.parseFromString(null, "text/html");
    }" did not throw
FAIL 'document.innerText = string' assigned via default policy (successful HTML transformation). assert_equals: expected "Quack, I want to be a duck!" but got "Hi, I want to be transformed!"
PASS 'document.innerText = null' assigned via default policy does not throw
